improve build time by 26s on my Macbook Pro (and more on Travis)
Apparently Swift type inference can’t handle adding more than 3 floats
in a single expression :|

before:
real	2m17.684s
user	2m31.372s
sys	0m27.125s

after:
real	1m51.157s
user	1m55.790s
sys	0m26.766s

+++ LayoutKitTests/InsetLayoutTests.swift
@@ -20,7 +20,7 @@ class InsetLayoutTests: XCTestCase {
         let arrangement = insetLabel.arrangement()
         AssertEqualDensity(arrangement.frame, [
             2.0: CGRect(x: 0, y: 0, width: 4+16.5+16, height: 2+20+8),
-            3.0: CGRect(x: 0, y: 0, width: 4+16+oneThird+16, height: 2+20-oneThird+8),
+            3.0: CGRect(x: 0, y: 0, width: CGFloat(4+16+oneThird+16), height: CGFloat(2+20-oneThird+8)),
         ])
         AssertEqualDensity(arrangement.sublayouts.first!.frame, [
             2.0:  CGRect(x: 4, y: 2, width: 16.5, height: 20),
